refactor(core): isolate propagation recovery
